requirements.txt
streamlit==1.20.0
groq==0.15.0
PyPDF2==2.10.0
python-docx==0.8.11
tiktoken==0.3.0
matplotlib==3.6.3
concurrent.futures; python_version<'3.9'
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0
google-auth==2.16.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
google-api-python-client==2.80.0
transformers==4.25.0
langchain==0.0.154
huggingface-hub==0.13.1
